refactor: move directories *-wagmi to *
