+++ src/store/login/saga.ts
@@ -1,10 +1,10 @@
 import { call, put, race, spawn, take, takeLatest } from 'redux-saga/effects';
 
 import { EmailLoginErrors, SagaActionTypes, Web3LoginErrors, reset, setErrors, setLoading, setStage } from '.';
-import { getSignedToken, getSignedTokenForConnector, isWeb3AccountConnected } from '../web3-wagmi/saga';
+import { getSignedToken, getSignedTokenForConnector, isWeb3AccountConnected } from '../web3/saga';
 import { authenticateByEmail, forceLogout, nonceOrAuthorize, terminate } from '../authentication/saga';
 import { Events as AuthEvents, getAuthChannel } from '../authentication/channels';
-import { Web3Events, getWeb3Channel } from '../web3-wagmi/channels';
+import { Web3Events, getWeb3Channel } from '../web3/channels';
 import { getHistory } from '../../lib/browser';
 
 export function* emailLogin(action) {
