+++ src/store/registration/saga.test.ts
@@ -34,7 +34,7 @@ import { fetchCurrentUser } from '../authentication/api';
 import { nonce as nonceApi } from '../authentication/api';
 import { throwError } from 'redux-saga-test-plan/providers';
 import { Connectors } from '../../lib/web3';
-import { getSignedTokenForConnector } from '../web3-wagmi/saga';
+import { getSignedTokenForConnector } from '../web3/saga';
 import { completeUserLogin } from '../authentication/saga';
 import { createConversation } from '../channels-list/saga';
 import { denormalize as denormalizeUser } from '../users';
