+++ src/store/saga.ts
@@ -1,6 +1,6 @@
 import { all, call, spawn } from 'redux-saga/effects';
 
-import { saga as web3Wagmi } from './web3-wagmi/saga';
+import { saga as web3Wagmi } from './web3/saga';
 import { saga as channelsList } from './channels-list/saga';
 import { saga as messages } from './messages/saga';
 import { saga as channels } from './channels/saga';
@@ -10,7 +10,7 @@ import { saga as theme } from './theme/saga';
 import { saga as createConversation } from './create-conversation/saga';
 import { saga as createInvitation } from './create-invitation/saga';
 import { saga as registration } from './registration/saga';
-import { saga as loginWagmi } from './login-wagmi/saga';
+import { saga as loginWagmi } from './login/saga';
 import { saga as rewards } from './rewards/saga';
 import { saga as pageLoad } from './page-load/saga';
 import { saga as editProfile } from './edit-profile/saga';
